requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
DateTime==4.3
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
dnspython==2.0.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Pillow==7.1.2
py3-validate-email==0.2.10
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.11
validate-email==1.3
whitenoise==5.2.0
zope.interface==5.1.0
